Move care charges controllers into common folder
