+++ LBH.AdultSocialCare.Api/V1/Controllers/Common/CareChargesControllers/PackageCareChargesController.cs
@@ -1,12 +1,12 @@
+using System;
+using System.Threading.Tasks;
 using Common.Exceptions.CustomExceptions;
 using LBH.AdultSocialCare.Api.V1.Boundary.Common.Response;
 using LBH.AdultSocialCare.Api.V1.UseCase.Common.Interfaces;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
-using System;
-using System.Threading.Tasks;
 
-namespace LBH.AdultSocialCare.Api.V1.Controllers.Common
+namespace LBH.AdultSocialCare.Api.V1.Controllers.Common.CareChargesControllers
 {
     [Route("api/v1/package-care-charges")]
     [ApiController]
